+++ test/ttlang/Translate/TTLToCpp/dma_loop_multi_tile_nontrivial_cb.mlir
@@ -75,7 +75,7 @@
 
 module {
   func.func @dma_loop_multi_tile(%arg0: tensor<64x64xf32, #layout_2x2>, %arg1: tensor<96x64xf32, #layout_3x2>)
-      attributes {ttl.base_cta_index = 2 : i64, ttl.crta_indices = [0, 1], ttl.kernel_thread = #ttkernel.thread<noc>} {
+      attributes {ttl.base_cta_index = 2 : i32, ttl.crta_indices = [0, 1], ttl.kernel_thread = #ttkernel.thread<noc>} {
     %c0 = arith.constant 0 : index
     %cb1 = ttl.bind_cb {cb_index = 0, buffer_factor = 2} : !ttl.cb<[2, 2], f32, 2>
     %cb2 = ttl.bind_cb {cb_index = 1, buffer_factor = 2} : !ttl.cb<[3, 1], f32, 2>
