+++ test/ttlang/Translate/TTLToCpp/dma_multi_tile_read.mlir
@@ -35,7 +35,7 @@
 // CHECK:   return;
 // CHECK-NEXT: }
 module {
-  func.func @dma_multi_tile_read(%arg0: tensor<64x64xf32, #layout>) attributes {ttl.base_cta_index = 1 : i64, ttl.crta_indices = [0], ttl.kernel_thread = #ttkernel.thread<noc>} {
+  func.func @dma_multi_tile_read(%arg0: tensor<64x64xf32, #layout>) attributes {ttl.base_cta_index = 1 : i32, ttl.crta_indices = [0], ttl.kernel_thread = #ttkernel.thread<noc>} {
     %c0 = arith.constant 0 : index
     %cb = ttl.bind_cb {cb_index = 0, buffer_factor = 2} : !ttl.cb<[1, 1], f32, 2>
     %xf = ttl.copy %arg0, %cb : (tensor<64x64xf32, #layout>, !ttl.cb<[1, 1], f32, 2>) -> !ttl.transfer_handle<read>
